fix(tests): apply camelCase/snake_case corrections across Accounting, Ecommerce, CRM, and HR modules
Corrected field naming inconsistencies in 40 test files:
- Accounting (24 files): AccountBalance, AccountMapping, AuditLog, ExchangeRatePolicy, FiscalPeriod, IdempotencyKey, JournalEntry, JournalLine tests
- Ecommerce (12 files): CheckoutSession model/seeder, tests for Coupon, InventoryReservation, PaymentTransaction, ProductComparison, ShippingMethod, ShoppingCart, WishlistItem
- CRM (1 file): PipelineStages ShowTest
- HR (2 files): Department and Position ShowTests

All corrections follow JSON:API standard (camelCase for API, snake_case for DB).
Removed 4 temporary test runner scripts (run_accounting_*.sh, run_billing_*.sh).

+++ Modules/Ecommerce/app/Models/CheckoutSession.php
@@ -77,7 +77,7 @@ class CheckoutSession extends Model
             return null;
         }
 
-        return now()->diffInMinutes($this->expires_at);
+        return (int) now()->diffInMinutes($this->expires_at);
     }
 
     // Business Logic Methods
